updated and increased size

+++ Projects/javaBasics/Reversenumber.java
@@ -5,11 +5,11 @@ public class Reversenumber
     {
         Scanner sc =new Scanner(System.in);
         System.out.println("Enter your number");
-        int num= sc.nextInt();
-        int rev =0;
+        long  num= sc.nextLong();
+        long rev =0;
         while (num!=0)
         {
-          int digit= num%10;
+          long digit= num%10;
           rev = rev*10 + digit;
           num = num/10;
         }
